Split mock HTTP tests into separate files per artifact type
Move lambda tests to lambda_artifact_test.go and ECS tests to
ecs_image_test.go, mirroring the production code file structure.
